app/requirements.txt
Flask==1.0.2
flask-cors
Flask-SQLAlchemy
PyMySQL
python-dotenv
